add support for task-defined options to ULTIMATE tool info module
For internal evaluations, we use task definition files that specify
ULTIMATE parameters pertaining to one particular program, e.g. custom
specifications (e.g preconditions) or assumptions about the program.

Specifying these in the task definition is useful, as the expected verdict
may differ depending on these parameters, and we can easily have multiple
task definition files with different parameters and verdicts for a program.

If an option is already present, we do not insert it again (or ULTIMATE
would fail). Thus, options specified in the benchexec XML take precedence,
which allows the same task definition files to be evaluated with multiple
run definitions.